Use packer
Packer has better features than paq and is more actively developed.

[fix #58 #60]
