+++ home/.config/nvim.symlink/.gitignore
@@ -1,1 +1,2 @@
 /plugged
+plugin/packer_compiled.lua
