+++ home/.config/nvim.symlink/scripts/package.json
@@ -1,0 +1,4 @@
+{
+  "$schema": "https://json.schemastore.org/package.json",
+  "type": "module"
+}
