+++ home/.config/nvim.symlink/tests/test_init.vim
@@ -1,4 +1,5 @@
 set rtp+=.
+set loadplugins
 let &rtp= &rtp . ',' . expand('<sfile>:h')
 runtime plugin/plenary.vim
